integrate app hosting with cli
